자유게시판 소팅, 헤드, 전반적 수정

+++ DoranDoran/src/main/java/dr/board/action/DeleteAction.java
@@ -16,6 +16,7 @@ public class DeleteAction implements Action{
 		
 		HttpSession session = request.getSession();
 		Integer user_num = (Integer)session.getAttribute("user_num");
+		Integer user_auth = (Integer)session.getAttribute("user_auth"); 
 		if(user_num==null) { //로그인이 되지 않은 경우
 			return "redirect:/member/loginForm.do";
 		}
@@ -25,9 +26,12 @@ public class DeleteAction implements Action{
 		BoardVO db_board = dao.getBoard(board_num);
 		if(user_num != db_board.getMem_num()) {
 			//로그인한 회원번호와 작성자 회원번호가 불일치
-			return "WPB-INF/views/common/notice.jsp";
+			return "WEB-INF/views/common/notice.jsp";
 		}
-		
+		if(user_num != db_board.getMem_num() && user_auth != 3) {
+			//로그인한 회원번호와 작성자 회원번호가 불일치
+			return "WEB-INF/views/common/notice.jsp";
+		}
 		//로그인한 회원번호와 작성자 회원번호가 일치
 		dao.deleteBoard(board_num);
 		//파일 삭제
